Added an animated GIF image suggestive of the "Organelle" demo
The file "organelle_fill.gif" contains a series of slices through a
synthetic "cell" containing two organelles. All objects are simple
spheres, and the GIF image contains the series of "slices" through the
model which generate the images. The thought is that Cell Organizer might
be able to generate geometry for CellBlender/MCell from such a simple
model. This proposed model is just a starting point.