A little more organizing of the README.md files.
The README.md files can point directly to other README.md files, but
then the files in the repository are not visible. Since GitHub has a
default behavior of showing the README.md file whenever a directory is
viewed, this commit changes the MCell link from a link to the README.md
file to a link to the mcell directory itself. If this works, then the
README.md file will be displayed along with the repository files. This
might be a more convenient way to navigate the site. If it works, the
other subdirectories will be handled the same way.